Refactor OpenBSD-like module into OpenBSD
